requirements.txt
mysql-connector-python==8.2.0
flask==2.3.3
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
